reactor: generate from templates

+++ project_scaffold/render.py
@@ -4,10 +4,14 @@ Description: Omit
 LastEditors: Rustle Karl
 LastEditTime: 2022.02.10 14:54
 """
+import os.path
 from pathlib import Path
-from typing import Tuple
+from typing import List, Tuple
 
 from jinja2 import Template
+
+from .common import GENERATOR_HEADER, create_common_files, get_package
+from .templates import TEMPLATES_PATH
 
 
 def render_by_jinja2(content: str, *args, **kwargs):
@@ -15,31 +19,98 @@ def render_by_jinja2(content: str, *args, **kwargs):
 
 
 def render_by_format(content: str, *args, **kwargs):
-    return content.format(*args, **kwargs).strip() + "\n"
+    return (
+        content.strip()
+        .replace("{{", "{{{{")
+        .replace("}}", "}}}}")
+        .format(*args, **kwargs)
+        + "\n"
+    )
 
 
 def render_templates_recursively(
-    p: Path, dst: Path, include_suffixes: Tuple[str], src: Path = None, **kwargs
+    p: Path,
+    dst: Path,
+    special_paths: Tuple[str] = None,
+    include_suffixes: Tuple[str] = None,
+    src: Path = None,
+    render=render_by_jinja2,
+    **kwargs
 ):
-    q = dst / (p.relative_to(src or p))
+    src = src or p
+    q = dst / (p.relative_to(src))
 
     if p.is_dir():
         q.mkdir(parents=True, exist_ok=True)
         for e in p.iterdir():
-            render_templates_recursively(e, dst, include_suffixes, src or p, **kwargs)
+            render_templates_recursively(
+                e,
+                dst,
+                special_paths=special_paths,
+                render=render_by_format
+                if render != render_by_format
+                and special_paths
+                and any(
+                    map(
+                        lambda op: os.path.normpath(p) == os.path.normpath(src / op),
+                        special_paths,
+                    )
+                )
+                else render,
+                include_suffixes=include_suffixes,
+                src=src,
+                **kwargs
+            )
 
     elif p.is_file():
         suffixes = p.suffixes
 
         if len(suffixes) > 1:
             suffix = suffixes[0]
+            last_suffix = suffixes[-1]
 
-            if suffix not in include_suffixes:
+            if last_suffix == ".py":
+                qs = render(q.as_posix(), **kwargs).strip().partition(".")
+                q = Path(qs[0]) / qs[-1]
+                q.parent.mkdir(parents=True, exist_ok=True)
+
+            elif include_suffixes and suffix in include_suffixes:
+                qs = str(q).partition(suffix)
+                q = Path(qs[0] + qs[-1])
+            else:
                 return
 
-            qs = str(q).partition(suffix)
-            q = Path(qs[0] + qs[-1])
+        if q.exists():
+            return
 
         kwargs["GOLANG_PACKAGE"] = p.parent.stem
-        content = render_by_jinja2(p.open(encoding="utf-8").read(), **kwargs)
+        content = render(p.open(encoding="utf-8").read(), **kwargs)
         q.open("w", encoding="utf-8", newline="\n").write(content)
+
+
+def render_templates(
+    relpath,
+    special_paths: List[str] = None,
+    include_suffixes: Tuple[str] = None,
+    folders: List[str] = None,
+    **kwargs
+):
+    package, package_title, package_underscore = get_package()
+
+    render_templates_recursively(
+        TEMPLATES_PATH / relpath,
+        Path.cwd(),
+        special_paths=special_paths,
+        include_suffixes=include_suffixes,
+        **{
+            "PACKAGE_TITLE": package_title,
+            "APP_NAME": package_underscore,
+            "GOLANG_MODULE": package_underscore,
+            "PYPI_PACKAGE": package,
+            "PYTHON_MODULE": package_underscore,
+            "MAKEFILE_HEADER": GENERATOR_HEADER,
+            **kwargs,
+        }
+    )
+
+    create_common_files(folders)
